+++ sdk/mycelium_sdk/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.4.0"
+__version__ = "0.5.1"
 
 from mycelium_sdk.context import AgentContext, StellarNetwork, TxResult
 from mycelium_sdk.hive import HiveClient